Rename to <CGAL/functional.h>
